Finalize the v1.4-branch (#2056)
* Update README for 1.4 (#2047)

* Update references for 1.4

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Add a table for the common components

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Update K8s version

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Knative: Use istio's local gateway (#2048)

By default Knative's local gateway will use the istio-ingressgateway Pod
for configuring traffic and binding VirtualServices. This means that all
in-cluster traffic will need to also pass via the ingress gateway.

We've noticed 302s from Notebooks curl-ing InferenceServices. This was
because the AuthService is authenticating all requests that go through
the ingress gateway. But since we now send in-cluster requests via the
ingress gateway this means that AuthService will also check them.

To avoid the above we provide an overlay for making Knative's local
gateway to use Istio's local gateway and not the ingress one.

See: https://github.com/kubeflow/manifests/issues/1966

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Update components of kubeflow/kubeflow for 1.4 (#2055)

* Sync with the 1.4 kubeflow/kubeflow repo

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Update README for 1.4

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>